+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [Unreleased]
 
-none! :tada:
-
+### Added
+ - examples! :beers:
+ 
 
 ## 4.4.0